Finished work on pressure forces.
Now working even with cables. This commit to save an idea where the contour term in the pressure forces is computed in the cables loop, saving ~10% in computing time. However this gives a strange looking influence to the cables orientation, so it will be removed and the contour term computed in the faces loop where it belongs.

Signed-off-by: pilcru <pierre.cuvilliers@gmail.com>